requirements.txt
torch==2.1.0
torchvision
numpy
pandas
scipy
Pillow
tqdm
tensorboard
matplotlib
scikit-learn
einops
openpyxl
